// src/interfaces/contractor.interface.ts
export interface Contractor {
    id: number;
    contractor_code: string;
    contractor_name: string;
    contractor_service: string;
    contractor_address: string;
    contractor_ofc_phone: string;
    contact_person: string;
    contact_mobile: string;
    contact_home_phone: string;
    email: string;
    website: string;
    comment: string;
}